split off CatalogContext and BaseObject subclasses into their own files
